refactored click paint elements and move DOM helpers to builder module
